requirements.txt
et_xmlfile==2.0.0
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
platformdirs==4.3.6
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.6.0
shiboken6==6.8.0.2
six==1.16.0
tzdata==2024.2
wheel==0.45.1
